Create a placeShip() and getState() function that successfully pass tests
The placeShip() function places a ship in valid coordinates such that the ship's length does not exceed the bounds of the gameboard and each coordinate is not already occupied by another ship.

The getState() function returns a copy of the state variable (the current state of the gameboard) to prevent mutation outside of the module.