Create a receiveAttack() function that successfully passes tests
The receiveAttack() function updates the state of the gameboard depending on the given coordinate. If the coordinate is empty, the attack is a miss. If the coordinate contains a ship, the ship gets a hit and the state of the board is updated to indicate a hit at that coordinate.