Create a Player factory and an attack() function and refactor the Gameboard factory
This commit successfully passes the player tests that require the user to place an attack at a given coordinate. The Gameboard factory was refactored to return null if an attack was invalid (i.e., when the user attempts to attack an already hit or missed coordinate) to make the attack() function return null for an invalid move. This return value can be used in a future implementation to design the app's UI.