Implement the ability for the user to make the first move
This commit allows the user to make the first move and consequently updates both the visual design of the webpage and the flow of the game. The next step is to allow the computer to make a move and to repeat this sequence of events until the game is over, denoted by the initial pseudocode in the game module.