Update the CSS
This commit adds a default transparent border around the h2 elements to prevent movement of elements when the active player is changed after each move. It also changes the background color of hovered occupied box elements to a darker shade of the ship color rather than the default gameboard color as well as updates several class names.