Refactor the Player factory
This commit refactors the generateCoords() function to set the min (inclusive) and max (exclusive) values of the coordinates. Prior to this commit, the function would return an invalid coordinate that would throw an error when attempting to place an attack on the user's gameboard since the values were out of an array's bounds (e.g., ['0', 'b']). The function now also checks that the generated coordinates have not already been generated to prevent the computer from attacking the same coordinate twice.

This commit also creates a getLastMove() function that returns the last move made by the Player.