Implement the ability for the computer to place an attack after the user makes a move
This commit refactors the game module by placing the code previously inside of initializeGame() outside of it (module scope) so the variables are accessible in playRound(), fixing a scope issue. It also calls other methods such that after the user places an attack, the computer will follow up with an attack of its own 2 seconds later, updating the display of the UI and continuing the flow of the game.