Improve the UI of the gameboard when hovering over boxes that have already been clicked
This commit reverts the cursor from pointer to default as well as maintain the initial background color of boxes that the user has already selected. This way, when the user goes to make their next move, hovering over already-selected boxes will show the user that those boxes are not selectable.