Dynamically create and display the game over modal
This commit removes the static HTML elements of the game over modal and instead dynamically creates the modal via JS once a player's ships have all been sunk. When considering the static elements of the HTML, it was decided that the game over modal shouldn't exist until the game ends, thus converting it into a dynamic element.