Display a valid or invalid placement of a selected ship piece when hovering over the initial gameboard
This commit changes the background color of the boxes to denote a valid or invalid placement of a ship on the gameboard. If the ship's length exceeds the bounds of the gameboard at a certain box, the boxes turn red. Likewise, if the ship's length is within the bounds of the gameboard, the boxes turn green. Future implementations will consider the logic when the user places their ship by clicking on a box on the gameboard.