Implement the ability for the user to place a horizontal ship on their gameboard
This commit displays a green area on the gameboard when hovering over boxes that are valid positions for the user to place a ship and red if they are not. When clicking on a valid box, the boxes containing the ship's length remain green and the data of the user's input is sent to the game module.

Future implementations will include the ability for the user to rotate a ship, reset all placed ships, and allowing the player to place only one kind of ship per ship piece (max 5 ships on a gameboard).