Update the UI of the ship pieces after the user makes a selection
This commit places a checkmark over the ship piece that the user placed on the gameboard and disables the piece to prevent the user from selecting and placing the same ship onto the gameboard. It also delegates the actively selected piece to the next ship piece once the user makes a selection on the gameboard for ease of interactivity. The user still has the option to click on any available ship piece as they please.