Fix rotate modal UI
This commit fixes the issue of selecting the opposite ship direction when the user clicks on the current direction. In other words, if the current direction is set to 'horizontal,' clicking on 'horizontal' would set the direction to 'vertical,' but this commit fixes that issue.

This commit also checks if the rotate modal already exists in the DOM, and if so, showRotationOptions() won't recreate the modal/reset it.