Implement the start game modal and fix bug of accidental DOM manipulation of completed gameboards when game starts
This commit validates whether or not the user has placed all of their ships when attempting to start the game. If all ships have not been placed, a message appears requiring the user to go back and place all of their ships, else the game will start and display the rendered gameboards of both the user and the opponent.

Prior to this commit, there was a bug that disabled the opponent's gameboard, which prevented the user from placing an attack. This was due to the DOM manipulation logic in the initial-gameboard module that query selected all of the boxes present on the DOM rather than just the initial gameboard. Specifying the query selectors fixed the bug and the user can now interact with the opponent's gameboard.